create or change item UX

+++ src/app/change-item/change-item.component.ts
@@ -15,6 +15,8 @@ export class ChangeItemComponent {
   @Input() item!: foodObject;
   @Output() itemChanged = new EventEmitter();
   @Output() itemDeleted = new EventEmitter();
+  @Input() confirmDeleteCount!: number;
+  @Output() confirmDeleteCountChange = new EventEmitter<number>();
   url =
     'https://my-list-a7fb0-default-rtdb.europe-west1.firebasedatabase.app/items';
 
@@ -34,10 +36,16 @@ export class ChangeItemComponent {
   }
 
   onDeleteItem() {
+    // debugger;
+    this.confirmDeleteCount += 1;
+    this.confirmDeleteCountChange.emit(this.confirmDeleteCount);
+    if (this.confirmDeleteCount < 2) return;
+
     this.authService.itemDeletedID = this.item.id!;
 
     this.foodData.deleteSingleItem(this.url).subscribe(() => {
       this.foodData.listItems.splice(this.item.itemId, 1);
+      this.confirmDeleteCount = 0;
       this.foodData.setUniqueId();
       this.itemDeleted.emit();
     });
